+++ jsonpath/function_extensions/__init__.py
@@ -1,3 +1,7 @@
+from .keys import keys
 from .length import length
 
-__all__ = ("length",)
+__all__ = (
+    "keys",
+    "length",
+)