+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Added support for function extensions.
 - Added the built-in `length()` function.
+- Added the built-in `count()` function. `count()` is an alias for `length()`
+- Added the built-in `keys()` function.
 
 ## Version 0.2.0